refactor(math): provide default solve methods on DerivativeSolver
This pull request lifts the shared `solve` and `solve_bracketed` logic into the `DerivativeSolver` trait as default methods, removing the duplicated implementations from each concrete solver. Implementors now only need to supply `config` and `refine`.

**Trait changes:**
* Added a `config` method to `DerivativeSolver` so the default `solve`/`solve_bracketed` implementations can access evaluation caps and domain bounds.
* Promoted `refine` to a required trait method, documenting the bracket invariants guaranteed by the driver.
* Moved the accuracy validation, auto-bracketing, and bracket-validation flow into the trait's default `solve` and `solve_bracketed` methods.

**Solver simplification:**
* Updated `Newton` and `NewtonSafe` to implement only `config` and `refine`, dropping their now-redundant `solve` and `solve_bracketed` overrides.
* Trimmed the imports in both solvers to drop the bracketing helpers that are no longer referenced directly.
