+++ crates/itofin/src/math/solvers1d/newton.rs
@@ -11,10 +11,7 @@
 
 use crate::errors::QlResult;
 use crate::fail;
-use crate::math::solver1d::{
-    Bracketed, DerivativeSolver, Function1D, Solver1DState, SolverConfig, bracket_by_stepping,
-    bracket_given,
-};
+use crate::math::solver1d::{DerivativeSolver, Function1D, Solver1DState, SolverConfig};
 use crate::types::Real;
 
 /// Newton-Raphson root finder.
@@ -48,8 +45,19 @@ impl Newton {
         self.config.upper_bound = Some(upper_bound);
         self
     }
+}
 
-    /// Newton iteration on a prepared bracket.
+impl Default for Newton {
+    fn default() -> Self {
+        Newton::new()
+    }
+}
+
+impl DerivativeSolver for Newton {
+    fn config(&self) -> &SolverConfig {
+        &self.config
+    }
+
     fn refine<G: Function1D>(
         &self,
         g: &mut G,
@@ -93,53 +101,6 @@ impl Newton {
             "maximum number of function evaluations ({}) exceeded",
             self.config.max_evaluations
         )
-    }
-}
-
-impl Default for Newton {
-    fn default() -> Self {
-        Newton::new()
-    }
-}
-
-impl DerivativeSolver for Newton {
-    fn solve<G: Function1D>(
-        &self,
-        mut g: G,
-        accuracy: Real,
-        guess: Real,
-        step: Real,
-    ) -> QlResult<Real> {
-        if accuracy <= 0.0 {
-            fail!("accuracy ({accuracy}) must be positive");
-        }
-        let accuracy = accuracy.max(Real::EPSILON);
-        // Bind before matching so the value-closure's borrow of `g` is released
-        // before `refine` takes `&mut g`.
-        let bracketed = bracket_by_stepping(&self.config, &mut |x| g.value(x), guess, step)?;
-        match bracketed {
-            Bracketed::Root(x) => Ok(x),
-            Bracketed::Ready(st) => self.refine(&mut g, accuracy, st),
-        }
-    }
-
-    fn solve_bracketed<G: Function1D>(
-        &self,
-        mut g: G,
-        accuracy: Real,
-        guess: Real,
-        x_min: Real,
-        x_max: Real,
-    ) -> QlResult<Real> {
-        if accuracy <= 0.0 {
-            fail!("accuracy ({accuracy}) must be positive");
-        }
-        let accuracy = accuracy.max(Real::EPSILON);
-        let bracketed = bracket_given(&self.config, &mut |x| g.value(x), guess, x_min, x_max)?;
-        match bracketed {
-            Bracketed::Root(x) => Ok(x),
-            Bracketed::Ready(st) => self.refine(&mut g, accuracy, st),
-        }
     }
 }
